service-auth-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2